Extract TypedDocumentBuilder from SyntheticDocumentBuilder
Move typed class generation (EmitTypedRow, EmitTypedRowCollection,
EmitTypedTable, AppendUsings) and table type name dictionary building
into a shared TypedDocumentBuilder in Analysis/. SyntheticDocumentBuilder
delegates to it for both Build() and BuildForDiagnostics().

This makes the typed compilation context reusable by the upcoming
SemanticAnalysisService without duplicating code.
